Move search to package, create view for detail
