+++ src/main/java/com/metaxiii/escalade/controller/path/search/Search.java
@@ -1,4 +1,4 @@
-package com.metaxiii.escalade.controller.path;
+package com.metaxiii.escalade.controller.path.search;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
